Several updates to workspace functionality when logging in and out
When a user logs in, it now uses their last workspace and channel (saved in DB)
to determine which workspace and channel to log the user into on sign in.

If that workspace or channel does not exist, the user is redirected to a default
workspace or channel.

Logging out now correctly removes a user from the workspace active user count, using
a combination of socketio logout event and redirect to the /logout flask route.