Hovering on messages in chat panel now displays relevant options to client
Now when a client hovers over a message it will display either edit
and delete options if it is their own message, or a private message link
if it is another user's message. Edit delete and private message functionality
to be added.

Index.hs refactored to better segregate various function types for readability